finsihed authorizarion jwt

+++ src/tasks/tasks.service.ts
@@ -1,6 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { create } from 'eslint/lib/rules/*';
+import { User } from 'src/auth/user.entity';
 import { CreateTaskDTO } from './dto/create-task.dto';
 import { GetTaskFilterDto } from './dto/get-task-filter.dto';
 import { StatusTask } from './task-status.enum';
@@ -17,11 +18,11 @@ export class TasksService {
 
     }
 
-    getTasks(filterDTO: GetTaskFilterDto): Promise<Task[]>{
+    getTasks(filterDTO: GetTaskFilterDto, user:User): Promise<Task[]>{
         // if(filterDTO.status ==='OPEN'){
         //     throw new Error('asdasdasd');
         // }
-        return this.taskRepository.getTasks(filterDTO);
+        return this.taskRepository.getTasks(filterDTO,user);
     }
 
     // getTasksWithFilters(taskfilterDto: GetTaskFilterDto): Task[] {
@@ -38,8 +39,8 @@ export class TasksService {
 
     //     return tasks;
     // }
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({where: {id, userId: user.id}});
 
         if (!found) {
             throw new NotFoundException(`Task Id ${id} not found`);
@@ -48,11 +49,11 @@ export class TasksService {
 
     }
 
-    async createTask(createTaskDto: CreateTaskDTO): Promise<Task>{
+    async createTask(createTaskDto: CreateTaskDTO, user: User): Promise<Task>{
     
-        return this.taskRepository.createTask(createTaskDto);
+        return this.taskRepository.createTask(createTaskDto, user);
     }
-    async deleteTask(id:number) {
+    async deleteTask(id:number, user:User) {
         
         // const found =  await this.getTaskById(id);
         // const obejectDeleted = await this.taskRepository.remove(found);
@@ -60,14 +61,16 @@ export class TasksService {
         // if(!obejectDeleted){
         //     throw new NotFoundException(`${id} not found`);
         // }
-        const result = await this.taskRepository.delete(id);
+        const result = await this.taskRepository.delete({id, userId: user.id});
         if(result.affected === 0){
             throw new NotFoundException(`${id} not found`);
         }
     }
     
-    async updateStatus(id:number, newStatus: StatusTask): Promise<Task> {
-        const task = await this.getTaskById(id);       
+    async updateStatus(id:number, 
+        newStatus: StatusTask,
+        user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);       
         task.status = newStatus;       
         return await this.taskRepository.save(task);
     }
